% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dates}
\alias{dates}
\title{A Date Function}
\usage{
dates(img)
}
\arguments{
\item{img}{this is the image}
}
\description{
This function allows you to express your love of cats.
}
\examples{
dates()
}
\keyword{date}
